+++ src/muzero_gojax/models/value.py
@@ -36,6 +36,6 @@ class TrompTaylorValue(base.BaseGoModel):
 
     def __call__(self, embeds):
         turns = gojax.get_turns(embeds)
-        sizes = gojax.compute_area_sizes(embeds).astype('int16')
+        sizes = gojax.compute_area_sizes(embeds).astype('bfloat16')
         n_idcs = jnp.arange(len(embeds))
         return sizes[n_idcs, turns.astype('uint8')] - sizes[n_idcs, (~turns).astype('uint8')]
